Add standards-compliant random-access (array) iterator
The iterator simply wraps a raw array. If T is the value type, the template is parameterized only on T*. It is not also parameterized by array size so that the same iterator template may be used even with dynamically-sized arrays (for example, vectors). For simplicity, the iterator's ctor also does not accept a size parameter, making the iterator completely unchecked. A possible improvement is to add a size (or other parameter) to the ctor so that operations can be checked.